Custom Background Rework
I am at least 90% sure this works how it is intended to. Will do some more testing.

The idea was to allow custom backgrounds without needing to completely blind the player by reworking how they are handled.

This did involve slightly altering the VR headset to make "None" be no background and "VR" being the Synthwave.

The idea is that any custom backgrounds should always have a type that isn't "None" for it. Since "None" is the default setting when getting the custom background.