Update Validation.js
Re-added support for the Restrain Property. As while it is not used by items anymore. Having it allows those that use console to enhance their experience to do so in a way that doesn't risk harming people.

As it currently is everything that could be done before can still be done, but doing so is much more risky and likely to break something. While yes this only effects those that use console, it allows greater diversity in RP oppotunities. Such as being able to hide things, set poses etc. Once again I would like to say that all of this is still possible, but doing so is now MUCH more likely to cause negative effects/outcomes.

I also removed the new validation of the lock member number since non number locks are used in the game currently. Off the top of my head the Nursery uses text instead of member numbers.