Optimised & Tested
Did some tests and all seems to be working right.

I swapped over the if statements to switch statements to make it more efficient. Also split each level into it's own function for easier editing. Doing the same for the baby talk function. I was considering adding a way to manually trigger baby talk without the item but felt that would need it's own pull request.

I THINK I managed to include everything that was initally talked about. When I tried submitting this before. Sorry for the long wait. I started it and had most of it written up, but then just bombed at the testing phase. Should work as it normally does but better. Also made it so that caps aren't instantly deleted when you are gagged. As I did before, though the higher levels those being 8+ aka the total gag levels do not have caps as it doesn't fit at that level.