Priority Update
A tiny change took actually a minute. I changed it so that the Strait Leotard has similar behaviour to the Futuristic Leotard.

That being that the suit itself is really low priority while the arms have a higher priority allowing items to be layered on top of the suit but below the arms.